# docs/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: 'SoloMe - Project Management App For Solopreneurs'
  description: 'The go-to project management app for solopreneurs who want to keep things simple, with no unnecessary features, just the essential tools you need to stay organized.'
  contact:
    name: Aminul Islam Abid
    email: aminul.i.abid@gmail.com
servers:
  - description: Dev Server
    url: http://localhost:3000/api/v1
  - description: Prod Server
    url: https://solome.vercel.app/api/v1
...
paths:
...
  /tasks:
    get:
      summary: Get all tasks
      description: Retrieve a list of all tasks for the authenticated user.
      security:
        - BearerAuth: []
      tags:
        - Tasks
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
          description: Filter tasks by priority.
        - name: labelId
          in: query
          schema:
            type: string
          description: Filter tasks by label.
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in-progress, completed]
          description: Filter tasks by status.
        - name: due_date
          in: query
          schema:
            type: string
            format: date
          description: Filter tasks by due date.
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Retrieve all tasks successfully.
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a new task
      description: Allows a user to create a new task with various attributes.
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Task created successfully.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{taskId}:
    patch:
      summary: Update a task
      description: Allows a user to update an existing task.
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
            example: '123e4567-e89b-12d3-a456-426614174000'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Task updated successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a task
      description: Allows a user to delete an existing task.
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: string
            example: '123e4567-e89b-12d3-a456-426614174000'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Task deleted successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          example: Write API Documentation
        content:
          type: string
          example: Write detailed API documentation for the SoloMe project.
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
          example: HIGH
        status:
          type: string
          enum: [TODO, IN_PROGRESS, COMPLETED]
          example: IN_PROGRESS
        dueDate:
          type: string
          format: date-time
        timeLog:
          type: object
          properties:
            startTime:
              type: string
              format: date-time
            endTime:
              type: string
              format: date-time
            duration:
              type: string
              example: 00:15:45
        subTasks:
          type: array
          items:
            type: object
            properties:
              task:
                type: string
                example: thing to do
              isComplete:
                type: boolean
        attachments:
          type: array
          items:
            type: string
          example:
            - https://example.com/file1.png
            - https://example.com/file2.pdf
        comments:
          type: array
          items:
            type: string
          example:
            - first comment
            - second comment
    Task:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: 123e4567-e89b-12d3-a456-426614174000
            projectId:
              type: string
              example: 123e4567-e89b-12d3-a456-426614174000
            labelId:
              type: string
              example: 123e4567-e89b-12d3-a456-426614174000
            title:
              type: string
              example: Write API Documentation
            description:
              type: string
              example: Write detailed API documentation for the SoloMe project.
            priority:
              type: string
              enum: [LOW, MEDIUM, HIGH]
              example: HIGH
            status:
              type: string
              enum: [TODO, IN_PROGRESS, COMPLETED]
              example: IN_PROGRESS
            dueDate:
              type: string
              format: date-time
            timeLog:
              type: object
              properties:
                startTime:
                  type: string
                  format: date-time
                endTime:
                  type: string
                  format: date-time
                duration:
                  type: string
                  example: 00:15:45
            subtask:
              type: array
              items:
                type: object
                properties:
                  task:
                    type: string
                    example: thing to do
                  isComplete:
                    type: boolean
            attachments:
              type: array
              items:
                type: string
              example:
                - https://example.com/file1.png
                - https://example.com/file2.pdf
            comments:
              type: array
              items:
                type: string
              example:
                - first comment
                - second comment
        - $ref: '#/components/schemas/Timestamp'
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Bad Request
              code:
                type: integer
                format: int32
                example: 400
              message:
                type: string
                example: "'field_name' field is required."
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
              code:
                type: integer
                format: int32
                example: 401
              message:
                type: string
                example: Invalid 'resource'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: 'error'
              code:
                type: integer
                format: int32
                example: 404
              message:
                type: string
                example: "'resource' not found."
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT